always import modules instead of classes
and move TestExtractor into its own module to be re-used for other tests

give test_class a more descriptive name
