+++ fex/__init__.py
@@ -1,4 +1,4 @@
 """Expose FeatureExtractor directly on pex module and config logging."""
 
 from fex.feature_extractor import FeatureExtractor # NOQA
-from fex.collection import FeatureExtractorCollection # NOQA
+from fex.collection import Collection # NOQA
